+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryPreDisposeHookRepository.java
@@ -188,7 +188,7 @@ final class Arez_RepositoryPreDisposeHookRepository extends RepositoryPreDispose
   @Override
   public final int hashCode() {
     if ( Arez.areNativeComponentsEnabled() ) {
-      return Double.hashCode( $$arezi$$_id() );
+      return Integer.hashCode( $$arezi$$_id() );
     } else {
       return super.hashCode();
     }
