+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithMultipleCtorsRepository.java
@@ -219,7 +219,7 @@ final class Arez_RepositoryWithMultipleCtorsRepository extends RepositoryWithMul
   @Override
   public final int hashCode() {
     if ( Arez.areNativeComponentsEnabled() ) {
-      return Double.hashCode( $$arezi$$_id() );
+      return Integer.hashCode( $$arezi$$_id() );
     } else {
       return super.hashCode();
     }
